Whisper_v3_large/requirements.txt
torch
transformers
datasets
accelerate
evaluate
jiwer
soundfile
librosa
ctranslate2
insanely-fast-whisper
pyyaml